TPermissionsConfigurationBehavior also needed to zap its manager
Moved the common properties of the PermissionManager in behaviors to. a trait.
